// src/app/core/shared-component/table/table.types.ts
export type KebabOption = {
    isEdit?: boolean;
    isDelete?: boolean;
    isDetail?: boolean;
    isExport?: boolean;
    isReject?: boolean;
    isAmount?: boolean;
    isApprove?: boolean;
    isPublish?: boolean;
    isDataTrustee?: boolean;
    isMenuAccessibility?: boolean;
    isRescheduled?: boolean;
    isDeleteDraft?: boolean;
    isEditDraft?: boolean;
    isCancelRequest?: boolean;
    isDownload?: boolean;
    isCreateInstallment?: boolean;
    isSetFinalApprover?: boolean;
    isActive?: boolean;
    isInactive?: boolean;
    isEditPending?: boolean;
    isEditUnpaid?: boolean;
    isDetailOpen?: boolean;
    isReAssign?: boolean;
    isViewLearningEvents?: boolean;
    isResetDevice?: boolean;
    isCreateRevision?: boolean;
    isAssignToPic?: boolean;
    isDuplicate?: boolean;
    isKeyIn?: boolean;
};

export type ColumnMap = {
    label: string;
    key?: string;
    isChip?: boolean;
    isCenter?: boolean;
    isAmount?: boolean;
    isFormula?: boolean;
    isCheckmark?: boolean;
    isUseClickRow?: boolean;
    isDateTimezone?: boolean;
    isUseClickCell?: boolean;
    isDateAndTimeT?: boolean;
    isTimeNoSecond?: boolean;
    isPublishStatus?: boolean;
    isSeparator?: boolean;
    separator?: string;
    keys?: [string, string];
    isDate?: boolean;
    isDateAttendCorrec?: boolean;
    isTime?: boolean;
    isRequired?: boolean;
    isList?: boolean;
    isManyDate?: boolean;
};